requirements.txt
readability-lxml
docopt
fake_useragent
scrapy-redis
#ScrapyElasticSearch
